Add support for navigating the buffer and correcting mistakes
Left and right arrow keys and backspace now move the cursor correctly and cause
the buffer to scroll. There's no deleting of characters, you can only overwrite
previously entered data. This was done to make the code simpler (no copying big
sections of the buffer around or using data structures to make it fast)

This has required fairly significant refactoring and changing of the state
variables. There turned out to be a lot of code specific to the hex editor. I
think the editor code will actually be simpler with a text editor.