Replace my own printing loop with a single print string BIOS call
It didn't really turn out as drastically simpler as I hoped because setting up
the arguments got a little ugly.

I was thinking I wanted to print lines like this as I add features like
insertion & deletion, and new lines. Now I'm not sure because we're going to do
string scanning to find the '\n' anyway and maybe we could print during that,
although I suppose it is better to not be calling as many interrupts.

Really, doing the memory mapped version would be best. Maybe there's a way to
call the BIOS to get the pointer.