Add support for new lines at the end of the buffer with line wrapping
Still in replacement mode. This is really just an intermediate step so I can
commit something working and test my line scanning code.

End goal is to add single line horizontal scrolling and switch to insert mode
with a gap buffer. This lets us never have to redraw more than a single line at
a time and avoids copying large parts of the buffer often when scrolling or
inserting.