Implemented line wrapping partially The lines don't reset when you scroll past like they should yet
Also added a little debug printing routine to the bootsector that fills the
space in the upper margin

I wrote this over a pretty long period, I had a buggy version done sometime in
February then world events caused me to not work on this for a while. Over the
last couple days I fixed up the bugs and got it to this point.