Start using a gap buffer for storage but still in replacement mode
Now I'm ready to start changing the UI behavior to insert mode. The idea
is that we maintain a single garbage data gap where the cursor is so
that we can shrink it one byte when we type to avoid copying the tail
data as often.

In both replacement mode and insert mode move_left and move_right will
just copy one character across the gap. Also both modes will have very
similar line-wrapping behavior (to avoid repainting the whole screen on
inserts and deletes). So I get to re-use all this code in insert mode.

Currently the gap never shrinks until we hit USER_CODE_MAX, and then it
stays shrunken because we don't really need a gap in replacement mode
(this code is just a transition state).

Also now the user code is null-terminated (had to stop using a globally
saved register for the end pointer)