Implement gap buffer expansion. Now insert mode is fully working.
Copied 2 bytes at a time to unroll the loop a bit and take advantage of
the register size. Used something like duff's device to handle odd
number character counts.

Also added a warning message for when we run out of buffer space because
we naturally have a good place to check in maybe_reset_gap. Plus we
needed some guard checks to stop us when the buffer is full since we
can't just replace chars now in the middle.