Relocate the stack and user code so they don't conflict with the editor code
Previously the stack was overlapping with the editor code, but it has
never actually been deep enough to overwrite anything. So I put the
stack up in the limited block of memory and moved the user code after
the editor code.

I also had to switch to unsigned comparison jumps (which I should have
been using the whole time) because now USER_CODE_MAX is greater than the
highest positive twos-compliment number.

I thought this was the bug I was seeing when I added the backspace code,
but turns out it wasn't. This is still a good idea though and I'm glad I
noticed.