Clean up the behavior when we hit USER_CODE_MAX
Now we propery set and clear the warning message in all cases. Also had to
ensure that we always have at least one char of gap space for move_right and
move_left to work right (need to use it to print a space sometimes).

Lastly made the gap expansion (from the middle of a line) push the gap out all
the way to the end even if it can't get the full GAP_SIZE. If we didn't, we
wouldn't be able to fully fill up the buffer from the middle.