Start parsing the arguments, add support for NO_ARG only right now
Factored out skip_to_end_of_line (for skipping comments) and skip_spaces
so that I could reuse them in the parse_arguments function.

Going to need extra return flags for the argument types probably. Still
not sure what to do about filling in the relative jumps and how I want
to decide if it's going to be near or short.

This was mostly written last night but it was left buggy.