Created a basic assembler tester binary to compare nasm output with mine
It runs in qemu because that's the easiest way to run this 8086 code. I
wanted to run it with a C wrapper using virtual-8086 mode but there's no
documentation really for the linux vm86 call so that was going to be a
lot of work.

It actually worked out really nicely this way. I think using both incbin
and %include of the test file is a neat trick.

Also
  - Made a directory for all the assembler files
  - Split the assembler data tables into a separate file
  - Fixed the boot script not working with files in folders