Convert the data table to array of struct instead of struct of array
This is to prepare for the symbol table which will have this same
data structure and will re-use the search_data_table call.

Also:
  - Fixed a bug where not all memory dereference types were supported
    correctly for `mem/reg, imm` instructions (we weren't calling
    write_mem_reg, it was old copy-pasted code). Added a test case for
    this.
  - No more ugly hack for the jmp extra_opcode!
  - Reworked some register allocations
  - Deleted some unnecessary "byte" and "word" specifiers

I re-organized the data table with a bunch of wild vim macros. Hopefully
I won't need to change up the data structure again. The tester program
worked great, I wouldn't have been able to do this refactor and keep it
bug-free without it.