Implemented writing the data for the symbol table
Looks like it works and is ready to use in the next step (but it doesn't
support local labels yet).

 - Included a pretty long debug routine to print out the symbol table
   instead of continuing to assemble.
 - Added a print_hex_byte call to the bootloader
 - Made the assembler tester support reading code that doesn't start at
   0 because now the symbol table starts at 0 and we need to skip it
   - Not super happy about how this turned out...

I wrote the code to write the symbol table data a while back but stopped
when I realized that the buffer isn't simply null-terminated (so
shift_data got more complicated). Today I wrote shift_data and the debug
printing routine for the symbol table and fixed up the tester after
debugging everything.