Fill label addresses in the symbol table as we see them also return the address or the unresolved status from parse_expression
We still need a final pass to fill in all the addresses for the
unresolved labels (the ones where the label is defined after the
instruction using it). To do that we'll have to clean up how we use the
stack in the top-level assembler function, so that we can store the list
of unresolved addresses on the stack.

This will return absolute addresses from parse_expression and insert
them in the code if the use was after the label definition. This is only
correct for some contexts but others need relative addresses, so this
commit does generate bad code sometimes.

I wrote this code a while ago (I don't remember now) and then I stopped
when I got frustrated by a giant errors list when I was having a bad
day. Yesterday I fixed up the errors, put in the strings, and double
checked the code.