Created a script to run bootable code in qemu and attach gdb
This was a huge pain to get working. I needed a work-around for qemu
forcing the architecture to 32 bit assembly and a helper script to make
debugging bootable code livable.

I spent a while trying to get debugging symbols to work. It almost did.
The strategy was from stackoverflow and it is to have nasm generate an
elf object file with debug info and then build an elf binary for gdb and
a flat binary for qemu. That all worked in gdb but the problem is making
an elf binary also moved around all of the code and I couldn't get the
flat binary from the obj file to correctly run in qemu because of that,
then if I used my normal binary with it (`nasm -f bin`) the addresses
wouldn't match.

Nevertheless I did get the debugger to work and it is just barely
manageable to do useful work with it even without the symbols.