The lisp interpreter datastructure works! The first steps
It makes a stack of pairs of words, the first is the type and the
second is the value. The lists are linked lists made of cons pairs which
are two pointers to nodes pushed onto the stack, which are used as prev
and next pointers with more cons pairs. Everything in the interpreter is
built out of this structure.

So far it builds the initial global env and then prints the global env.
I still need to implement parsing and eval and fill in the builtins, but
all of those things are built from the pieces that work now.