First steps getting eval working
 - Implemented lookup_env
 - Apply only works for primitive functions, lisp defined functions are
   harder
 - eval is complete already, but eval_each is still needed.
 - Because I made two different ATOM types and didn't copy the strings
   into the lisp memory, and the strings are null terminated or strings
   with a count, I have to write up to 4 different string compare
   functions. Kind of annoying. It's tempting to just swap the arguments
   to re-use the one csatom vs atom function and just write atom vs
   atom.

Just a little bit more to go, most of the hard parts are done now. After
I get the apply function completed (which needs eval_each) I also need
to implement the rest of the primitives, but hopefull that is
straightforward.

I wrote most of this on Sunday and then debugged it today and yesterday.
I didn't realize I needed csatom vs csatom equals until today.

This took a lot of debugging again. The code is getting very tight given
the registers that I have made global, it's getting hard to setup
arguments for function calls without a lot of stack operations.