Fix the prim_car and prim_cdr functions
I misunderstood how it works and I evaluated at the end when I changed
from calling eval_each to just calling eval on the first arg. It needs
to be evaluated before calling car so you call car on the evaluated
result.

It should be (car '(foo bar)) not (car ('foo 'bar))