Refactor the lisp code to use cx for the environment var not the stack
This saves a lot of pushing and popping and copying the env pointer
around. We didn't really need the type part of the argument so I saved a
word. Using a register we can actually forward arguments to the next
function seemlessly, which doesn't work with the stack because the
return address gets in the way.