Complete the minimal lisp interpreter!
It is Turing complete now! Like LISP 1 there are no native numbers so
you would have to represent them as lists.

I decided to add the pair? primitive instead of atom? like in McCarthy's
paper because I liked the idea from scheme that all of the type
predicates should be mutually exclusive and atom? would have to return
true for numbers and atoms later in order to be the predicate that you
use to avoid calling car/cdr with a non-CONS type which is an error.

There's no apply/eval primitive which some say is necessary, but I think
it isn't necessary fundamentally because you could just implement lisp
in lisp using the primitives provided. Of course it would not be as
efficient and you would have to reimplement the primitives I already
made in assembly but it would work theoretically.

Also I realized that most of this lisp variant is a lot like scheme.
I kinda wanted to rename #t to TRUE but #t is what scheme uses so I'll
keep it for compatibility.

There were assembly bugs in eval_each and apply that were uncovered in
the process of doing this.