Reorganize the memory map to turn lisp into a true REPL
Now everything has a separate 64k segment and we can preserve the lisp
stack between runs.

The big refactoring part was swapping all references of ds and es in
bootstrap-lisp.asm. It was also tricky that I needed to preserve the
user input code in order for the pointers in the lisp stack to be valid
since for ATOM strings we point into the code.

This changed the interface for run_code by a lot and now I will have to
refactor the assembler code as well to match the new API.