Fix bugs from changing scan_forward and adding [cs:user_code_start]
I guess I was actually using the fact that bp was saved in scan_forward
and I didn't notice because it was an unspoken parameter to print_line.
While now it's up to the caller and pretty much in every case I do have
to reset bp I think the calls to print_line are clearer now and I
actually did get to use the un-saved bp in a few places to simplify the
code. So overall I do still think it's better this way.

There were also some places I was still comparing with 0 or 1 for the
beginning of the buffer and instead we need to use [cs:user_code_start]
now for the lisp repl.