Implement the up arrow key in the text editor
It took a while to get this right. At first I wanted to make it so when
you move up, it tries it's best to keep the same distance from the left
newline on the line above, like vim does. This created a lot of cases
the editor state could be in that I had to deal with.

So I ended up changing the design of how the up arrow key works to match
the design of the editor and the simplest way it can work give the
VGA API. Since we have single line scrolling (because horizontally
scrolling the whole screen would mean a full repaint) it seems to make
more sense to make the editor instead try it's best to keep the cursor
column visually the same instead of position in the buffer the same.

This way when you press up it will always go to the place you can see on
the screen exactly one line up, unless the above line is shorter in
which case it will go to the end.

I didn't really want to have to save global state for the shorter lines
like vim does anyway. I actually had written an entire routine to
repaint the top line scrolled last night and ended up deleting it today.

Now that I have it figured out implementing down in the same way should
be straightforward.