Properly check for errors in lisp parsing
Also fix not null terminating the output when the lisp interpreter is
passed only whitespace.

Also fix the text-editor sample run_code implementation after the
refactor.

I still need to handle errors during eval better. The primitives can
sometimes return errors. Not sure if the new error type is for parsing
only or if we also want to use it in eval errors.