Split up the lisp code and start a lisp tester program
The lisp/tester.asm code is very ugly right now but I just got it mostly
working and I wanted to commit all of these changes. It needs a lot of
improvements to the output strings and code cleanups still.

I have this setup so that I can run the tests with qemu in headless mode
on github actions and then check if it passed or failed.

I also had to split out the lisp code into a library you can include and
call into. This ended up adding one function call overhead to the repl
loop but I don't think it's a big deal.