I realized eq? wasn't buggy, fix the tests
I checked tinylisp and this is how it acts. eq? can only check if the
exact object is the same, or the atom is the same value. So if you want
to check if the structure of two lists are the same you need to write a
function using eq? to compare each atom.