Add a lot of interesting lisp code to the debug text, improve comments
I was exploring if macro is necessary to have in lisp or if you could do
what macros do without it being a builtin. Because I was able to find a
way to write and/or with shortcutting without macros by using quote and
eval, I now think that macros are not fundamental.

I do think it will be nice to have macros later but I want this first
release to be extremely minimal and not have any primitives that aren't
necessary for a complete lisp. For one thing macros kind of fit in with
the "bootstrap" idea.

I do have an idea that will allow users to run code on un-evaluated
lists but it will be more difficult and interesting than simple lisp
macros :)