Implement the eval primitive
I resisted adding this for a long time because I wasn't convinced that
it is fundamental to have in a complete lisp. I did initially find some
situations where it seemed using eval or apply is the most natural
thing, but it wasn't always clear that you must use it and there's no
other way.

Later I realized that since you can easily implement lisp in lisp you
can actually implement eval and apply within lisp and use the built-in
list datastructures for the interpreter memory. Although this would be
much more inefficient than the built-in eval you would be able to
execute lisp code that way. However I eventually realized that if you
did this you would not be able to run any of the built-in primitives
within your lisp in lisp, you would have to re-implement them.
Additionally because you don't have the real eval, you would effectively
be running everything quoted; your interpreter would be a layer deeper
where everything is quoted and if you wanted to implement lisp in lisp
in lisp everything would be double quoted.

So ultimately I realized eval (aka unquote) is actually necessary to
have, especially when doing code-gen as if you were writing a macro.

The funny thing is, it's only a few lines of assembly to add this
primitive after all that thought about it.

P.S. For most of the time I believed that if I added anything it should
be apply not eval because apply is more simple. However once I heard
that eval can also be thought of as unquote I was convinced I needed
eval instead. This way it's also possible to do a global env lookup of
an atom without me also providing an assoc primitive.