Commit to cond using != nil comparisons instead of == #t
I think the != 0 tradition from C is actually a neat way to introduce
some syntax shortcuts, and lisp needs all the shortcuts it can get. If I
changed this a lot of code that uses cond would have to be written
differently so that things return #t more.

All the tests pass now!