Don't print anything for define expressions
Improve a lot of comments, resolving some TODOs and adding some others.

Also made a minor code change to remove a single add instruction that I
had left a TODO for (I think there's more like this but it's not really
worth it).

I also thought through some of the error checking in eval_each and made
a minor adjustment to the type checking. I also thought through and left
comments where it's necessary to check for stack overflows for x86 and
for the lisp stack (ignoring overflows in parsing and printing for now).