Set the overscan color to light blue!
This doesn't show up in qemu but when I boot on my modern x86 machine
there is a black border shown by default and this allows you to set that
color.

Overscan had meaning for CRT monitors where it wasn't possible to always
place the screen image exactly so there were borders to compensate.

It was very difficult to figure out the way to set the color value
correctly for this. It's a different format than in the int 0x10
printing routines. The IBM manuals weren't very helpful in explaining
it, maybe it's in there and I just don't know where to look. I found a
PC programming tutorial book and that explained that setting the
secondary color bits all 1 is the same as the intensity bit in the other
format.